# shared/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Real Assessment Platform API
  description: |
    REST API for the Real Assessment Platform -- an AI-powered candidate
    evaluation system that analyzes resumes, conducts interviews, and
    generates comprehensive scoring reports.
  version: "1.0.0"
  contact:
    name: API Support
    email: api-support@assessment.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.assessment.example.com/api/v1
    description: Production
  - url: https://staging-api.assessment.example.com/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (admin only)
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, interviewer, candidate]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
...
  /resumes:
    post:
      tags: [Resumes]
      summary: Upload a resume for parsing
      operationId: uploadResume
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Resume file (PDF, DOCX, TXT)
                user_id:
                  type: string
                  description: Associate with specific user (admin only)
      responses:
        "201":
          description: Resume uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [Resumes]
      summary: List resumes (with optional filters)
      operationId: listResumes
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: user_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, parsed, error]
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeList"
  /resumes/{resume_id}:
    get:
      tags: [Resumes]
      summary: Get resume details including parsed data
      operationId: getResume
      security:
        - BearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Resume with parsed data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetail"
...
  /interviews:
    post:
      tags: [Interviews]
      summary: Create a new interview session
      operationId: createInterview
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewRequest"
      responses:
        "201":
          description: Interview created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"

    get:
      tags: [Interviews]
      summary: List interviews
      operationId: listInterviews
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: candidate_id
          in: query
          schema:
            type: string
        - name: interviewer_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewList"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, interviewer, candidate]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        email_verified:
          type: boolean
...
    UserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # -- Resumes --
    Resume:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
        content_type:
          type: string
        status:
          type: string
          enum: [pending, parsed, error]
        created_at:
          type: string
          format: date-time
    ResumeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resume"
        total:
          type: integer
    ResumeDetail:
      allOf:
        - $ref: "#/components/schemas/Resume"
        - type: object
          properties:
            parsed_data:
              $ref: "#/components/schemas/ParsedResume"
    ParsedResume:
      type: object
      properties:
        skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              category:
                type: string
              proficiency:
                type: string
        experience:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              title:
                type: string
              start_date:
                type: string
              end_date:
                type: string
              description:
                type: string
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
              degree:
                type: string
              field:
                type: string
              graduation_year:
                type: integer
...
    CreateInterviewRequest:
      type: object
      required: [candidate_id, interviewer_id, scheduled_at]
      properties:
        candidate_id:
          type: string
          format: uuid
        interviewer_id:
          type: string
          format: uuid
        scheduled_at:
          type: string
          format: date-time
        questions:
          type: array
          items:
            type: string
        notes:
          type: string
    Interview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        candidate_id:
          type: string
          format: uuid
        interviewer_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        scheduled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    InterviewList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Interview"
        total:
          type: integer
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object